requirements.txt
# This file was autogenerated via `uv export`.
-e .
a-unet==0.0.16
accelerate==0.34.2
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3 ; python_full_version < '3.11'
attrs==24.2.0
audio-diffusion-pytorch==0.1.3
biopython==1.84
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6 ; sys_platform == 'win32' or platform_system == 'Windows'
contourpy==1.3.0
cycler==0.12.1
data-science-types==0.2.23
datasets==2.21.0
dill==0.3.8
distlib==0.3.8
docker-pycreds==0.4.0
einops==0.8.0
evo-model==0.1.2
exceptiongroup==1.2.2 ; python_full_version < '3.11'
filelock==3.15.4
flash-attn==2.6.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
gitpython==3.1.43
huggingface-hub==0.24.6
identify==2.6.0
idna==3.8
iniconfig==2.0.0
inquirerpy==0.3.4
jinja2==3.1.4
joblib==1.4.2
kan-gpt @ git+https://github.com/hssn-20/kan-gpt@9eabbb29efcebcf446393e45cb5bcd13975241fc
kiwisolver==1.4.7
levenshtein==0.25.1
markupsafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
nodeenv==1.9.1
numpy==2.1.1
nvidia-cublas-cu12==12.1.3.1 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-cuda-cupti-cu12==12.1.105 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-cuda-nvrtc-cu12==12.1.105 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-cuda-runtime-cu12==12.1.105 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-cudnn-cu12==9.1.0.70 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-cufft-cu12==11.0.2.54 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-curand-cu12==10.3.2.106 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-cusolver-cu12==11.4.5.107 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-cusparse-cu12==12.1.0.106 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-nccl-cu12==2.20.5 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-nvjitlink-cu12==12.6.68 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-nvtx-cu12==12.1.105 ; platform_machine == 'x86_64' and platform_system == 'Linux'
packaging==24.1
pandas==2.2.2
pfzy==0.3.4
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.8.0
prompt-toolkit==3.0.47
protobuf==5.28.0
psutil==6.0.0
pyarrow==17.0.0
pydantic==2.9.0
pydantic-core==2.23.2
pyparsing==3.1.4
pytest==8.3.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywavelets==1.7.0
pyyaml==6.0.2
rapidfuzz==3.9.7
regex==2024.7.24
requests==2.32.3
ruff==0.6.4
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
sentry-sdk==2.13.0
setproctitle==1.3.3
setuptools==74.1.2
six==1.16.0
smmap==5.0.1
stripedhyena==0.2.2
sympy==1.13.2
threadpoolctl==3.5.0
tokenizers==0.19.1
tomli==2.0.1 ; python_full_version < '3.11'
torch==2.4.1
torchaudio==2.4.1
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0 ; python_full_version < '3.13' and platform_machine == 'x86_64' and platform_system == 'Linux'
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
wandb==0.17.8
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.9.11
